requirements.txt
Flask
pandas
scikit-learn
xgboost
joblib
requests
beautifulsoup4
python-whois